+++ package.json
@@ -87,6 +87,7 @@
     "prop-types": "^15.6.0",
     "proxy-polyfill": "^0.1.7",
     "react": "~16.9.0",
+    "react-autosuggest": "^10.1.0",
     "react-bootstrap": "~0.33.0",
     "react-codemirror2": "^6.0.0",
     "react-dom": "~16.9.0",
